+++ app/Http/Controllers/Admin/RasController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Ras;
+use App\Ternak;
 use App\DataTables\RasDataTable;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
@@ -127,6 +128,13 @@ class RasController extends Controller
     public function destroy($id)
     {
         $data = Ras::findOrFail($id);
-        $data->delete();
+
+        if(Ternak::where('ras_id', $id)->exists()){
+            $err = 'Data ras id '. $id .' tidak dapat dihapus.';
+            return response()->json(['error' => $err]);
+        }
+        else{
+            $data->delete();
+        }
     }
 }
